+++ packages/amplify_datastore/android/src/main/kotlin/com/amazonaws/amplify/amplify_datastore/types/model/FlutterModelSchema.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -15,35 +15,57 @@
 
 package com.amazonaws.amplify.amplify_datastore.types.model
 
+import com.amazonaws.amplify.amplify_core.cast
+import com.amplifyframework.core.model.Model
+import com.amplifyframework.core.model.ModelIndex
 import com.amplifyframework.core.model.ModelSchema
 import com.amplifyframework.core.model.SerializedModel
 
 data class FlutterModelSchema(val map: Map<String, Any>) {
     val name: String = map["name"] as String
     private val pluralName: String? = map["pluralName"] as String?
-    private val authRules: List<FlutterAuthRule>? =
-        (map["authRules"] as List<Map<String, Any>>?)?.map { FlutterAuthRule(it) }
-    private val fields: Map<String, FlutterModelField> =
-        (map["fields"] as Map<String, Any>).mapValues { FlutterModelField(it.value as Map<String, Any>) }
+    private val authRules: List<FlutterAuthRule> = when (val rawAuthRules = map["authRules"]) {
+        is List<*> -> (rawAuthRules.cast<Map<String, Any>>()).map { FlutterAuthRule(it) }
+        else -> emptyList()
+    }
+    private val fields: Map<String, FlutterModelField> = when (val rawFields = map["fields"]) {
+        is Map<*, *> -> (rawFields.cast<String, Map<String, Any>>()).mapValues { FlutterModelField(it.value) }
+        else -> emptyMap()
+    }
     private val associations: Map<String, FlutterModelAssociation> =
         fields.filterKeys { key -> fields[key]?.getModelAssociation() != null }
             .mapValues { it.value.getModelAssociation()!! }
 
+    // In amplify-android, primaryKey and indexes are categories as ModelIndex
+    private val indexes: Map<String, ModelIndex> = run {
+        val parsedIndexes = emptyMap<String, ModelIndex>().toMutableMap()
+        val rawIndexes = map["indexes"]
+
+        if (rawIndexes is List<*>) {
+            parsedIndexes.putAll(
+                (rawIndexes.cast<Map<String, Any>>()).associate {
+                    (it["name"] as String?
+                        ?: FlutterModelIndex.unnamedIndexKey) to FlutterModelIndex(it).convertToNativeModelIndex()
+                }.toMap()
+            )
+        }
+
+        parsedIndexes
+    }
+
     fun convertToNativeModelSchema(): ModelSchema {
-        var builder: ModelSchema.Builder = ModelSchema.builder()
+        return ModelSchema.builder()
             .name(name)
             .pluralName(pluralName)
             .fields(fields.mapValues { it.value.convertToNativeModelField() })
-
-        if (!authRules.isNullOrEmpty()) {
-            builder = builder.authRules(authRules.map { it.convertToNativeAuthRule() })
-        }
-        if (!associations.isNullOrEmpty()) {
-            builder = builder.associations(
-                associations.mapValues { it.value.convertToNativeModelAssociation() }
-            )
-        }
-        builder.modelClass(SerializedModel::class.java)
-        return builder.build()
+            .indexes(indexes)
+            .authRules(authRules.map { it.convertToNativeAuthRule() })
+            .associations(associations.mapValues { it.value.convertToNativeModelAssociation() })
+            .modelClass(SerializedModel::class.java)
+            .modelType(Model.Type.USER)
+            // to allow amplify-android correctly interpret custom primary key
+            // schema version needs to be >= 1
+            .version(1)
+            .build();
     }
 }