+++ packages/amplify_datastore/example/lib/models/BelongsToChildImplicit.dart
@@ -17,10 +17,10 @@
 // Generated files can be excluded from analysis in analysis_options.yaml
 // For more info, see: https://dart.dev/guides/language/analysis-options#excluding-code-from-analysis
 
-// ignore_for_file: public_member_api_docs, file_names, unnecessary_new, prefer_if_null_operators, prefer_const_constructors, slash_for_doc_comments, annotate_overrides, non_constant_identifier_names, unnecessary_string_interpolations, prefer_adjacent_string_concatenation, unnecessary_const, dead_code
+// ignore_for_file: public_member_api_docs, annotate_overrides, dead_code, dead_codepublic_member_api_docs, depend_on_referenced_packages, file_names, library_private_types_in_public_api, no_leading_underscores_for_library_prefixes, no_leading_underscores_for_local_identifiers, non_constant_identifier_names, null_check_on_nullable_type_parameter, prefer_adjacent_string_concatenation, prefer_const_constructors, prefer_if_null_operators, prefer_interpolation_to_compose_strings, slash_for_doc_comments, sort_child_properties_last, unnecessary_const, unnecessary_constructor_name, unnecessary_late, unnecessary_new, unnecessary_null_aware_assignments, unnecessary_nullable_for_final_variable_declarations, unnecessary_string_interpolations, use_build_context_synchronously
 
 import 'ModelProvider.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
+import 'package:amplify_core/amplify_core.dart';
 import 'package:flutter/foundation.dart';
 
 /// This is an auto generated class representing the BelongsToChildImplicit type in your schema.
@@ -36,9 +36,13 @@ class BelongsToChildImplicit extends Model {
   @override
   getInstanceType() => classType;
 
+  @Deprecated(
+      '[getId] is being deprecated in favor of custom primary key feature. Use getter [modelIdentifier] to get model identifier.')
   @override
-  String getId() {
-    return id;
+  String getId() => id;
+
+  BelongsToChildImplicitModelIdentifier get modelIdentifier {
+    return BelongsToChildImplicitModelIdentifier(id: id);
   }
 
   String? get name {
@@ -109,9 +113,9 @@ class BelongsToChildImplicit extends Model {
   }
 
   BelongsToChildImplicit copyWith(
-      {String? id, String? name, BelongsToParent? belongsToParent}) {
+      {String? name, BelongsToParent? belongsToParent}) {
     return BelongsToChildImplicit._internal(
-        id: id ?? this.id,
+        id: id,
         name: name ?? this.name,
         belongsToParent: belongsToParent ?? this.belongsToParent);
   }
@@ -138,8 +142,10 @@ class BelongsToChildImplicit extends Model {
         'updatedAt': _updatedAt?.format()
       };
 
-  static final QueryField ID =
-      QueryField(fieldName: "belongsToChildImplicit.id");
+  static final QueryModelIdentifier<BelongsToChildImplicitModelIdentifier>
+      MODEL_IDENTIFIER =
+      QueryModelIdentifier<BelongsToChildImplicitModelIdentifier>();
+  static final QueryField ID = QueryField(fieldName: "id");
   static final QueryField NAME = QueryField(fieldName: "name");
   static final QueryField BELONGSTOPARENT = QueryField(
       fieldName: "belongsToParent",
@@ -160,7 +166,7 @@ class BelongsToChildImplicit extends Model {
     modelSchemaDefinition.addField(ModelFieldDefinition.belongsTo(
         key: BelongsToChildImplicit.BELONGSTOPARENT,
         isRequired: false,
-        targetName: "belongsToChildImplicitBelongsToParentId",
+        targetNames: ["belongsToChildImplicitBelongsToParentId"],
         ofModelName: (BelongsToParent).toString()));
 
     modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
@@ -185,4 +191,42 @@ class _BelongsToChildImplicitModelType
   BelongsToChildImplicit fromJson(Map<String, dynamic> jsonData) {
     return BelongsToChildImplicit.fromJson(jsonData);
   }
+}
+
+/// This is an auto generated class representing the model identifier
+/// of [BelongsToChildImplicit] in your schema.
+@immutable
+class BelongsToChildImplicitModelIdentifier
+    implements ModelIdentifier<BelongsToChildImplicit> {
+  final String id;
+
+  /// Create an instance of BelongsToChildImplicitModelIdentifier using [id] the primary key.
+  const BelongsToChildImplicitModelIdentifier({required this.id});
+
+  @override
+  Map<String, dynamic> serializeAsMap() => (<String, dynamic>{'id': id});
+
+  @override
+  List<Map<String, dynamic>> serializeAsList() => serializeAsMap()
+      .entries
+      .map((entry) => (<String, dynamic>{entry.key: entry.value}))
+      .toList();
+
+  @override
+  String serializeAsString() => serializeAsMap().values.join('#');
+
+  @override
+  String toString() => 'BelongsToChildImplicitModelIdentifier(id: $id)';
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) {
+      return true;
+    }
+
+    return other is BelongsToChildImplicitModelIdentifier && id == other.id;
+  }
+
+  @override
+  int get hashCode => id.hashCode;
 }