+++ packages/amplify_datastore/ios/Classes/types/temporal/FlutterTemporal.swift
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -19,16 +19,16 @@ import Foundation
 
 // Stores an ISO 8601 String to be saved to Appsync
 struct FlutterTemporal: TemporalSpec {
-    
     let iso8601: String
     init(iso8601String: String) {
         self.iso8601 = iso8601String
         self.foundationDate = Date()
     }
+
     var iso8601String: String {
         iso8601
     }
-    
+
     // Not needed functions
     // In order to properly adhere to "TemporalSpec" these functions must be implemented
     // This class is solely for transmitting a ISO 8601 Date String to Appsync so these other functions are not needed
@@ -37,10 +37,12 @@ struct FlutterTemporal: TemporalSpec {
         self.iso8601 = ""
         self.foundationDate = date
     }
+
     static func now() -> FlutterTemporal {
         return FlutterTemporal(Foundation.Date())
     }
+
     func iso8601FormattedString(format: TemporalFormat, timeZone: TimeZone) -> String{
-        return "";
+        return ""
     }
 }