+++ packages/amplify_datastore/test/amplify_datastore_stream_controller_test.dart
@@ -20,7 +20,7 @@ import 'package:amplify_datastore/amplify_datastore_stream_controller.dart';
 import 'package:amplify_test/amplify_test.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'test_models/ModelProvider.dart';
+import 'package:amplify_test/test_models/ModelProvider.dart';
 
 void main() {
   const MethodChannel datastoreChannel =