+++ packages/amplify_test/lib/amplify_test.dart
@@ -19,9 +19,12 @@ export 'src/ambiguate.dart';
 
 /// Auth Utils
 export 'src/integration_test_utils/auth_cognito/integration_test_auth_utils.dart';
-export 'src/integration_test_utils/auth_cognito/types/admin_create_user_response.dart';
 export 'src/integration_test_utils/auth_cognito/types/confirm_sign_up_response.dart';
 export 'src/integration_test_utils/auth_cognito/types/delete_user_response.dart';
 
 /// Common Utils
 export 'src/json.dart';
+
+/// Stubs
+export 'src/stubs/amplify_auth_cognito_stub.dart';
+export 'src/stubs/amplify_stub.dart';