+++ packages/auth/amplify_auth_cognito/example/integration_test/get_current_user_test.dart
@@ -13,11 +13,10 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_api/amplify_api.dart';
+import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_test/amplify_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:amplify_flutter/amplify_flutter.dart';
 
 import 'utils/mock_data.dart';
 import 'utils/setup_utils.dart';
@@ -31,9 +30,7 @@ void main() {
 
   group('getCurrentUser', () {
     setUpAll(() async {
-      await configureAuth(additionalPlugins: [
-        AmplifyAPI(),
-      ]);
+      await configureAuth();
 
       // create one user for all tests
       await adminCreateUser(