+++ packages/auth/amplify_auth_cognito/example/integration_test/sign_in_sign_out_test.dart
@@ -13,24 +13,23 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_api/amplify_api.dart';
-import 'package:amplify_test/amplify_test.dart';
-import 'package:integration_test/integration_test.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_test/amplify_test.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
 
 import 'utils/mock_data.dart';
 import 'utils/setup_utils.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-  late String username;
-  late String password;
   group('signIn', () {
+    late String username;
+    late String password;
+
     setUp(() async {
-      await configureAuth(additionalPlugins: [
-        AmplifyAPI(),
-      ]);
+      await configureAuth();
 
       // create new user for each test
       username = generateUsername();
@@ -109,6 +108,9 @@ void main() {
     });
 
     testWidgets('should sign a user out', (WidgetTester tester) async {
+      final username = generateUsername();
+      final password = generatePassword();
+
       await adminCreateUser(
         username,
         password,