+++ packages/storage/amplify_storage_s3/example/tool/pull_test_backend.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+APP_ID=$STORAGE_APP_ID ../../../../build-support/pull_backend_by_app_id.sh