+++ packages/amplify/amplify_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.4-rc.1 (2022-07-20)
+
+- chore: bump amplify-android to v1.37.0-cpkey-preview.2
+
 ## 0.5.0 (2022-05-17)
 
 ### Breaking Changes